Allow double values in Attributes

+++ src/main/java/tamaized/regutil/RegUtil.java
@@ -895,7 +895,7 @@ public class RegUtil {
 		final UUID id;
 		final String type;
 
-		public ModAttribute(String name, float defaultValue, UUID id, String type) {
+		public ModAttribute(String name, double defaultValue, UUID id, String type) {
 			super(name, defaultValue);
 			this.id = id;
 			this.type = type;
